Fix react doctor route utility export warning
